+++ models/bag_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, Integer, String, func
+from sqlalchemy import Column, Integer, Numeric, String, func
 from sqlalchemy.dialects.postgresql import TIMESTAMP
 from sqlalchemy_serializer import SerializerMixin
 from models.base import Base
@@ -12,7 +12,7 @@ class BagModel(Base, SerializerMixin):
     video_url = Column(String, nullable=True)
     name = Column(String, nullable=False)
     description = Column(String, nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Numeric(10, 2), nullable=False)
 
     created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
     updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now(), onupdate=func.now())
