+++ models/shipping_method_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, Integer, String, func
+from sqlalchemy import Column, Integer, Numeric, String, func
 from sqlalchemy.dialects.postgresql import TIMESTAMP
 from sqlalchemy_serializer import SerializerMixin
 
@@ -10,8 +10,8 @@ class ShippingMethodModel(Base, SerializerMixin):
 
     id = Column(Integer, primary_key=True, unique=True, nullable=False)
     name = Column(String, nullable=False)
-    fee = Column(Float, nullable=False)
-    discount_fee = Column(Float, nullable=False)
+    fee = Column(Numeric(10, 2), nullable=False)
+    discount_fee = Column(Numeric(10, 2), nullable=False)
 
     created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
     updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now(), onupdate=func.now())
